Minor refactoring to FileAccessExecutor.
